+++ src/components/types/index.ts
@@ -1,17 +1,12 @@
 import React from 'react'
-import {
-  DeepPartial,
-  FieldValues,
-  FormContextValues,
-  Mode,
-} from 'react-hook-form'
+import { DeepPartial, FieldValues, Mode, UseFormMethods } from 'react-hook-form'
 
 import { JSONSchemaType, IDSchemaPair } from '../../JSONSchema'
 import { CustomValidators } from '../../hooks/validators'
 
 export interface JSONFormContextValues<
   FormValues extends FieldValues = FieldValues
-> extends FormContextValues<FormValues> {
+> extends UseFormMethods<FormValues> {
   schema: JSONSchemaType
   idMap: IDSchemaPair
   customValidators?: CustomValidators
@@ -27,8 +22,8 @@ export type OnSubmitType = (props: OnSubmitParameters) => void | Promise<void>
 export type FormContextProps<FormValues extends FieldValues = FieldValues> = {
   formProps?: Omit<React.HTMLAttributes<HTMLFormElement>, 'onSubmit'>
   validationMode?: Mode
-  revalidateMode?: Mode
-  submitFocusError?: boolean
+  revalidateMode?: Exclude<Mode, 'onTouched' | 'all'>
+  shouldFocusError?: boolean
   onChange?: (data: JSONSchemaType) => void
   onSubmit?: OnSubmitType
   noNativeValidate?: boolean